+++ src/App/VRChatContentManager.App/Pages/GettingStarted/GuideConnectUnityPage.axaml.cs
@@ -1,12 +1,10 @@
-﻿using Avalonia;
 using Avalonia.Controls;
-using Avalonia.Markup.Xaml;
 
 namespace VRChatContentManager.App.Pages.GettingStarted;
 
-public partial class GuideFinishPage : UserControl
+public partial class GuideConnectUnityPage : UserControl
 {
-    public GuideFinishPage()
+    public GuideConnectUnityPage()
     {
         InitializeComponent();
     }