+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Add Unity Setup Guide [`#110`](https://github.com/project-vrcz/content-manager/pull/110)
+  - Include install connect package, connect unity to app.
+  - You can directly jump to home page if you connect unity to app during guide.
+
 ## [1.2.0] - 2025-12-18
 
 ### Added